Fix to #3736 - Query :: some queries with Queryable.Cast<T> don't work
Problem was that during optimization, when we flatten subqueries we didn't account for Cast result operator properly. This would fail when trying to access a property on a derived type, when derived type was coming from a subquery with Cast result operator.

Fix is to wrap qsres to that (former) subquery with Convert to the last type that the subquery was casted into (in case multiple cast operators are used).

Also fixed another issue around this area:

- We tried to lift Distinct in some cases, which could lead to incorrect results. E.g

from outer in (ctx.Customers.Distinct())
select outer.Name

would get translated to:

(from outer in ctx.Customers
select outer.Name).Distinct()